Tomchris/config import export (#2)
* Updated parsing args

* Adding basic support for travis

* Added CLFs for branches, debug, local run, config files

* Corrected branch ref

* Moved variables and assignments around in wizard.sh to ensure CFL variable assignments are respected. Added examples.

* Corrected CFL and syntax issue

* testing travis

+++ wizard.sh
@@ -1,53 +1,122 @@
 #!/bin/bash
 
-wizard_basedir=~/.pf9-wizard
-wizard_venv=${wizard_basedir}/wizard-venv
-venv_python=${wizard_venv}/bin/python
-venv_activate=${wizard_venv}/bin/activate
-wizard_branch=master
-wizard_url=https://raw.githubusercontent.com/platform9/express-wizard/${wizard_branch}/wizard.py
-wizard_url_lib=https://raw.githubusercontent.com/platform9/express-wizard/${wizard_branch}/globals.py
-wizard_tmp_script=/tmp/pf9-wizard.py
-wizard_tmp_lib=/tmp/globals.py
-pip_url=https://bootstrap.pypa.io/get-pip.py
-pip_path=/tmp/get_pip.py
+####################################################################################################
+#
+# Example:
+#    Deploy and run wizard against master
+#         ./wizard.sh 
+#    Deploy and test wizard against master. Wizard will execute and then exit before the menu.
+#         ./wizard.sh 
+#    Use the current branch for the build and enable debugging
+#         ./wizard.sh -d -b=$(git branch 2> /dev/null | sed -e '/^[^*]/d' -e 's/* \(.*\)/\1/')
+#    Local deployment using the current branch debugging enabled (DEVELOPMENT ONLY!)
+#         ./wizard.sh -l -d -b=$(git branch 2> /dev/null | sed -e '/^[^*]/d' -e 's/* \(.*\)/\1/')
+#
+####################################################################################################
 
-# functions
-usage() {
-    echo "Usage: $(basename $0) [-i|--init]"
-    exit 1
-}
+start_time=$(date +"%s.%N")
 
 assert() {
-    if [ $# -gt 0 ]; then echo "ASSERT: ${1}"; fi
+    if [ $# -gt 0 ]; then echo "ASSERT: $(basename $0) : ${1}"; fi
     exit 1
 }
 
+debugging() {
+    if [[ ${debug_flag} ]]; then
+	echo "DEBUGGING: $(date +"%T") : $(bc <<<$(date +"%s.%N")-${start_time}) :$(basename $0) : ${1}"
+    fi
+}
+
+# parse commandline
+for i in "$@"; do
+  case $i in
+    -h|--help)
+	echo "Usage: $(basename $0)"
+	echo "	  [-i|--init]"
+	echo "	  [-t|--test]"
+	echo "	  [-l|--local]"
+	echo "	  [-d|--debug=]"
+	echo "	  [-c|--config=]"
+	echo "	  [-b|--branch=]"
+	echo ""
+	exit 0
+        shift
+        ;;
+    -b=*|--branch=*)
+	if [ -n ${i#*=} ]; then
+	    wizard_branch="${i#*=}"
+	else
+	    assert "Branch name must be provided";
+	fi
+	shift
+	;;
+    -c=*|--config=*)
+	config_file="${i#*=}"
+	if [ -z ${config_file} ];
+	    then assert "config file not provided: ${config_file}" 
+	elif [[ ! -r ${config_file} ]];
+	    then assert "failed to access config file: ${config_file}";
+	fi
+	shift
+	;;
+    -d|--debug)
+	debug_flag="${i#*=}"
+	shift
+        ;;
+    -d=*|--debug=*)
+	debug_flag="${i#*=}"
+	shift
+        ;;
+    -l|--local)
+        run_local="--local"
+	shift
+        ;;
+    -t|--test)
+        test_wizard="--test"
+	shift
+        ;;
+    -i|--init)
+        init_flag="--init"
+        shift
+        ;;
+    *)
+    echo "$i is not a valid command line option."
+    echo ""
+    echo "For help, please use $0 -h"
+    echo ""
+    exit 1
+    ;;
+    esac
+    shift
+done
+
 init_venv_python() {
-    if [ ${python_version} == 2 ]; then
+    if [[ ${python_version} == 2 ]]; then
         pyver="";
     else 
         pyver="3";
     fi
-    echo "Initializing Virtual Environment Python ${python_version}"
+    echo "Initializing Virtual Environment using Python ${python_version}"
+    #Validate and initialize virtualenv
     if [ "$(virtualenv --version -p python${pyver} > /dev/null 2>&1; echo $?)" -ne 0 ]; then
-        which pip > /dev/null 2>&1
-        if [ $? -ne 0 ]; then
+        #Validating pip
+	which pip > /dev/null 2>&1
+	if [ $? -ne 0 ]; then
             echo "ERROR: missing package: pip (attempting to install using get-pip.py)"
             curl -s -o ${pip_path} ${pip_url}
             if [ ! -r ${pip_path} ]; then assert "failed to download get-pip.py (from ${pip_url})"; fi
             python${pyver} ${pip_path}
             if [ $? -ne 0 ]; then
                 echo "ERROR: failed to install package: pip (attempting to install via 'sudo get-pip.py')"
-                sudo pythoni${pyver} ${pip_path} > /dev/null 2>&1
+                sudo python${pyver} ${pip_path} > /dev/null 2>&1
                 if [ $? -ne 0 ]; then
                     assert "Please install package: pip"
                 fi
             fi
         fi
-
-        echo "ERROR: missing python package: virtualenv (attempting to install via 'pip install virtualenv')"
-        pip${pyver} install virtualenv > /dev/null 2>&1
+	echo "ERROR: missing python package: virtualenv (attempting to install via 'pip install virtualenv')"
+        # Attemping to Install virtualenv
+	pip${pyver} install virtualenv > /dev/null 2>&1
         if [ $? -ne 0 ]; then
             echo "ERROR: failed to install python package (attempting to install via 'sudo pip install virtualenv')"
             sudo pip${pyver} install virtualenv > /dev/null 2>&1
@@ -63,23 +132,50 @@ init_venv_python() {
 
 ## main
 
-# parse commandline
-while [ $# -gt 0 ]; do
-    case ${1} in
-    -h|--help)
-        usage
-        ;;
-    -i|--init)
-        if [ -d ${wizard_basedir} ]; then
-            rm -rf ${wizard_basedir}
-            if [ -d ${wizard_basedir} ]; then assert "failed to remove ${wizard_basedir}"; fi
-        fi
-        ;;
-    esac
-    shift
-done
+if [ -z ${wizard_branch} ]; then wizard_branch=master; fi
+if [ -z ${run_local} ]; then
+    pf9_repo_dir=~/.pf9-wizard
+    wizard_script=/tmp/pf9-wizard.py
+    wizard_lib=/tmp/globals.py
+else
+    debugging "Using Local Install skipping Downloads"
+    pf9_repo_dir="$(dirname "$(readlink -fm "$0")")"
+    wizard_script=${pf9_repo_dir}/wizard.py
+    wizard_lib=${pf9_repo_dir}/globals.py
+fi
+
+wizard_basedir=~/.pf9-wizard
+wizard_venv=${wizard_basedir}/wizard-venv
+venv_python=${wizard_venv}/bin/python
+venv_activate=${wizard_venv}/bin/activate
+wizard_url=https://raw.githubusercontent.com/platform9/express-wizard/${wizard_branch}/wizard.py
+wizard_url_lib=https://raw.githubusercontent.com/platform9/express-wizard/${wizard_branch}/globals.py
+pip_url=https://bootstrap.pypa.io/get-pip.py
+pip_path=/tmp/get_pip.py
+
+# Merge runtime arguments
+if [[ -n ${wizard_branch} ]]; then args+=" --branch ${wizard_branch}";fi
+if [[ -n ${test_wizard} ]]; then args+=" ${test_wizard}";fi
+if [[ -n ${run_local} ]]; then args+=" ${run_local}";fi
+if [[ -n ${config_file} ]]; then args+=" --config ${config_file}";fi
+if [[ -n ${debug_flag} ]]; then 
+    if [[ ${debug_flag} == "-d" || ${debug_flag} == "--debug"  ]]; then
+	args+=" --debug 1"
+    else
+	args+=" --debug ${debug_flag}"
+    fi
+fi
+debugging "CLFs that will be passed to wizard.py:${args}"
+
 
 # initialize installation directory
+if [[ -n ${init_flag} ]]; then
+    debugging "DELETEING wizard_basedir: ${wizard_basedir}"
+    if [ -d ${wizard_basedir} ]; then
+        rm -rf ${wizard_basedir}
+        if [ -d ${wizard_basedir} ]; then assert "failed to remove ${wizard_basedir}"; fi
+    fi
+fi
 if [ ! -d ${wizard_basedir} ]; then
     mkdir -p ${wizard_basedir}
     if [ ! -d ${wizard_basedir} ]; then assert "failed to create directory: ${wizard_basedir}"; fi
@@ -90,9 +186,10 @@ which python > /dev/null 2>&1
 if [ $? -ne 0 ]; then assert "Python stack missing"; fi
 
 # configure python virtual environment
+debugging "Configuring virtualenv"
 if [ "$(ls -A ${wizard_venv} > /dev/null 2>&1; echo $?)" -ne 0 ]; then
     for ver in {3,2}; do #ensure python3 is first
-        if [ -x "$(which python${ver})" ]; then
+        if [ -x "$(which python${ver} > /dev/null 2>&1)" ]; then
 	    python_version="$(python${ver} <<< 'import sys; print(sys.version_info[0])')"
 	    break
         fi
@@ -102,53 +199,65 @@ else
     echo "INFO: using exising virtual environment"
 fi
 
-# remove cached files
-if [ -f ${wizard_tmp_script} ]; then
-    rm -f ${wizard_tmp_script}
-    if [ -f ${wizard_tmp_script} ]; then assert "failed to remove cached file: ${wizard_tmp_script}"; fi
-fi
-if [ -f ${wizard_tmp_lib} ]; then
-    rm -f ${wizard_tmp_lib}
-    if [ -f ${wizard_tmp_lib} ]; then assert "failed to remove cached file: ${wizard_tmp_lib}"; fi
+if [ -z ${run_local} ]; then
+    # remove cached files
+    if [ -f ${wizard_script} ]; then
+	debugging "Removing Temp file: ${wizard_script}"
+	rm -f ${wizard_script}
+	if [ -f ${wizard_script} ]; then assert "failed to remove cached file: ${wizard_tmp_script}"; fi
+    fi
+    if [ -f ${wizard_lib} ]; then
+        debugging "Removing Temp file: ${wizard_lib}"
+	rm -f ${wizard_lib}
+	if [ -f ${wizard_lib} ]; then assert "failed to remove cached file: ${wizard_tmp_lib}"; fi
+    fi
+
+    # download files
+    debugging "Download ${wizard_script} from: ${wizard_url}"
+    if [ "$(curl -s --fail -o ${wizard_script} ${wizard_url}; echo $?)" -ne 0 ]; then
+	assert "failed to download Platform9 Express Wizard (from ${wizard_url})"; fi
+    debugging "Download ${wizard_lib} from: ${wizard_url_lib}"
+    if [ "$(curl -s --fail -o ${wizard_lib} ${wizard_url_lib}; echo $?)" -ne 0 ]; then
+	assert "failed to download Platform9 Express Wizard Gobals(from ${wizard_url_lib})"; fi
 fi
 
-# download files
-if [ "$(curl -s --fail -o ${wizard_tmp_script} ${wizard_url}; echo $?)" -ne 0 ]; then
-    assert "failed to download Platform9 Express Wizard (from ${wizard_url})"; fi
-if [ "$(curl -s --fail -o ${wizard_tmp_lib} ${wizard_url_lib}; echo $?)" -ne 0 ]; then
-    assert "failed to download Platform9 Express Wizard Gobals(from ${wizard_tmp_lib})"; fi
-
+debugging "Upgrade pip"
 # upgrade pip
 (. ${venv_activate} && pip install pip --upgrade > /dev/null 2>&1)
 
 # install Openstack CLI
+debugging "Installing Openstack CLI"
 reqs=https://raw.githubusercontent.com/platform9/support-locker/master/openstack-clients/requirements.txt
-constraints=http://raw.githubusercontent.com/openstack/requirements/stable/pike/upper-constraints.txt
+constraints=https://raw.githubusercontent.com/openstack/requirements/stable/pike/upper-constraints.txt
 (. ${venv_activate} && pip install --upgrade --requirement ${reqs} --constraint ${constraints} > /dev/null 2>&1)
 
 # install Openstack SDK
+debugging "Install OpenstackSDK"
 (. ${venv_activate} && pip install openstacksdk==0.12 > /dev/null 2>&1)
+
+launch_wizard="(. ${venv_activate} && ${venv_python} ${wizard_script}${args})"
+debugging "Wizard launch command: ${launch_wizard}"
 
 # start pf9-wizard in virtual environment
 flag_started=0
 while [ ${flag_started} -eq 0 ]; do
-    (. ${venv_activate} && ${venv_python} ${wizard_tmp_script})
+    eval ${launch_wizard}
     if [ $? -eq 0 ]; then
         flag_started=1
     else
-        stdout=$(. ${venv_activate} && ${venv_python} ${wizard_tmp_script})
+        stdout=$(eval ${launch_wizard})
         echo "${stdout}" | grep "ASSERT: Failed to import python module:" > /dev/null 2>&1
         if [ $? -eq 0 ]; then
-            module_name=$(echo "${stdout}" | cut -d : -f3 | awk -F ' ' '{print $1}' | sed -e "s/'//g")
-            echo "--> attempting to installing missing module: [${module_name}]"
-            if [ "${python_version}" == "2" ]; then
-                (. ${venv_activate}; pip install ${module_name} > /dev/null 2>&1)
-            elif [ "${python_version}" == "3" ]; then
-                (. ${venv_activate}; python -m pip install ${module_name} > /dev/null 2>&1)
-            fi
-            if [ $? -ne 0 ]; then assert "failed to install missing module"; fi
+    	module_name=$(echo "${stdout}" | cut -d : -f3 | awk -F ' ' '{print $1}' | sed -e "s/'//g")
+    	echo "--> attempting to installing missing module: [${module_name}]"
+    	if [ "${python_version}" == "2" ]; then
+    	    (. ${venv_activate}; pip install ${module_name} > /dev/null 2>&1)
+    	elif [ "${python_version}" == "3" ]; then
+    	    (. ${venv_activate}; pip3 install ${module_name} > /dev/null 2>&1)
+    	fi
+    	if [ $? -ne 0 ]; then assert "failed to install missing module"; fi
         else
-            assert "${stdout}"
+    	assert "${stdout}"
         fi
     fi
 done
